+++ qabelbox/src/androidTest/java/de/qabel/qabelbox/ui/views/FolderChooserActivityTest.kt
@@ -35,10 +35,10 @@ class FolderChooserActivityTest {
     val defaultIntent: Intent
         get() =  Intent(InstrumentationRegistry.getTargetContext(), FolderChooserActivity::class.java).apply {
             putExtra(ACTIVE_IDENTITY, identity.keyIdentifier)
+            putExtra(FolderChooserActivity.TEST_RUN, true)
         }
 
     @Before
-    @Throws(Throwable::class)
     fun setUp() {
         val mContext = InstrumentationRegistry.getTargetContext()