+++ qabelbox/src/main/java/de/qabel/qabelbox/box/views/FileUploadView.kt
@@ -1,0 +1,16 @@
+package de.qabel.qabelbox.box.views
+
+import de.qabel.box.storage.dto.BoxPath
+import de.qabel.qabelbox.box.presenters.FileUploadPresenter
+import de.qabel.qabelbox.box.provider.DocumentId
+
+interface FileUploadView {
+
+    var identity: FileUploadPresenter.IdentitySelection
+    var path: BoxPath.FolderLike
+    var filename: String
+
+    fun startUpload(documentId: DocumentId)
+
+}
+