+++ qabelbox/src/main/java/de/qabel/qabelbox/identity/interactor/IdentityInteractor.kt
@@ -1,16 +1,13 @@
 package de.qabel.qabelbox.identity.interactor
 
-import de.qabel.core.config.Identities
 import de.qabel.core.config.Identity
 import de.qabel.core.drop.DropURL
 import de.qabel.core.repository.exception.EntityExistsException
 import rx.Single
 import java.io.FileDescriptor
 
-interface IdentityUseCase {
+interface IdentityInteractor : ReadOnlyIdentityInteractor {
 
-    fun getIdentity(keyId: String): Single<Identity>
-    fun getIdentities(): Single<Identities>
 
     fun createIdentity(alias: String, dropUrl: DropURL, prefix: String,
                        email: String, phone: String): Single<Identity>